feat: refactor install wizard and settings app, update routing and HTML links
