+++ src/amplifier_distro/server/apps/settings/static/settings.html
@@ -668,7 +668,7 @@ const state = {
   loaded: false,
 };
 
-const API = '/apps/install-wizard';
+const API = '/apps/settings';
 
 /* ================================================================== */
 /*  API Helpers                                                        */
