+++ src/amplifier_distro/server/apps/slack/static/slack-setup.html
@@ -439,7 +439,7 @@ body {
   </div>
 
   <div class="footer">
-    <a href="/apps/install-wizard/settings">Back to settings</a>
+    <a href="/apps/settings/">Back to settings</a>
   </div>
 </div>
