+++ src/amplifier_distro/server/apps/web_chat/static/index.html
@@ -372,7 +372,7 @@ body {
     </div>
   </div>
   <div class="header-right">
-    <a href="/apps/install-wizard/settings" class="settings-btn">
+    <a href="/apps/settings/" class="settings-btn">
       <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
         <circle cx="12" cy="12" r="3"/><path d="M12 1v2M12 21v2M4.22 4.22l1.42 1.42M18.36 18.36l1.42 1.42M1 12h2M21 12h2M4.22 19.78l1.42-1.42M18.36 5.64l1.42-1.42"/>
       </svg>
@@ -551,7 +551,7 @@ sendBtn.addEventListener('click', sendMessage);
     '<p><strong>Getting started:</strong></p>' +
     '<ul>' +
       '<li>Use the <code>amplifier</code> CLI for full-featured terminal sessions</li>' +
-      '<li>Manage providers and features in <a href="/apps/install-wizard/settings">Settings</a></li>' +
+      '<li>Manage providers and features in <a href="/apps/settings/">Settings</a></li>' +
       '<li>Sessions created via the Bridge API will connect here</li>' +
     '</ul>' +
     '<p><strong>Learn more:</strong></p>' +
